Added Images to Force Textures Dropdown
