+++ src/egui_extensions/mod.rs
@@ -1,0 +1,1 @@
+pub mod selectable_value_with_image;
